requirements.txt
# Assistant MJ - Les Lames du Cardinal
# Requirements

# Core
streamlit>=1.30.0
pyyaml>=6.0

# LangChain
langchain>=0.1.0
langchain-community>=0.0.20
langchain-ollama>=0.1.0

# RAG & Embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# PDF extraction (choose one or both)
pdfplumber>=0.10.0  # Recommended
PyPDF2>=3.0.0       # Fallback

# Data visualization
plotly>=5.18.0
pandas>=2.0.0

# Optional but recommended
torch>=2.0.0        # For GPU acceleration of embeddings
# Use: pip install torch --index-url https://download.pytorch.org/whl/cu118  (for CUDA 11.8)
# Or: pip install torch --index-url https://download.pytorch.org/whl/cpu     (for CPU only)